extract and combine all disparate action creators
