+++ applications/desktop/src/notebook/epics/comm.js
@@ -4,16 +4,13 @@ import { merge } from "rxjs/observable/merge";
 import { map, retry, switchMap } from "rxjs/operators";
 import { ofType } from "redux-observable";
 
+import { commOpenAction, commMessageAction } from "@nteract/core/actions";
+
 import { createMessage, ofMessageType, childOf } from "@nteract/messaging";
 
 import type { ActionsObservable } from "redux-observable";
 
-import {
-  COMM_OPEN,
-  COMM_MESSAGE,
-  COMM_ERROR,
-  NEW_KERNEL
-} from "@nteract/core/actionTypes";
+import { NEW_KERNEL } from "@nteract/core/actionTypes";
 
 /**
  * creates a comm open message
@@ -69,58 +66,6 @@ export function createCommCloseMessage(
     content: { comm_id, data },
     parent_header
   });
-}
-
-/**
- * creates a comm error action
- * @param  {error} error any type of error to pass on
- * @return {Object}       Flux standard error action
- */
-export const createCommErrorAction = (error: Error) =>
-  of({
-    type: COMM_ERROR,
-    payload: error,
-    error: true
-  });
-
-/**
- * Action creator for comm_open messages
- * @param  {jmp.Message} a comm_open message
- * @return {Object}      COMM_OPEN action
- */
-export function commOpenAction(message: any) {
-  // invariant: expects a comm_open message
-  return {
-    type: COMM_OPEN,
-    data: message.content.data,
-    metadata: message.content.metadata,
-    comm_id: message.content.comm_id,
-    target_name: message.content.target_name,
-    target_module: message.content.target_module,
-    // Pass through the buffers
-    buffers: message.blob || message.buffers
-    // NOTE: Naming inconsistent between jupyter notebook and jmp
-    //       see https://github.com/n-riesco/jmp/issues/14
-    //       We just expect either one
-  };
-}
-
-/**
- * Action creator for comm_msg messages
- * @param  {jmp.Message} a comm_msg
- * @return {Object}      COMM_MESSAGE action
- */
-export function commMessageAction(message: any) {
-  return {
-    type: COMM_MESSAGE,
-    comm_id: message.content.comm_id,
-    data: message.content.data,
-    // Pass through the buffers
-    buffers: message.blob || message.buffers
-    // NOTE: Naming inconsistent between jupyter notebook and jmp
-    //       see https://github.com/n-riesco/jmp/issues/14
-    //       We just expect either one
-  };
 }
 
 /**
